Material shouldn't be platform specific
