+++ platform/OpenGL/CMakeLists.txt
@@ -13,8 +13,6 @@ target_sources(${LIBRARY_TARGET_NAME} PRIVATE
         Voxymore/OpenGL/OpenGLRenderAPI.hpp
         Voxymore/OpenGL/OpenGLTexture2D.cpp
         Voxymore/OpenGL/OpenGLTexture2D.hpp
-        Voxymore/OpenGL/OpenGLMaterial.cpp
-        Voxymore/OpenGL/OpenGLMaterial.hpp
         )
 
 target_include_directories(${LIBRARY_TARGET_NAME} PRIVATE
